app/backend/requirements.txt
asgiref==3.8.1
black==24.10.0
click==8.1.7
Django==4.2.16
flake8==7.1.1
isort==5.13.2
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.6
psycopg2-binary==2.9.10
pycodestyle==2.12.1
pyflakes==3.2.0
sqlparse==0.5.2
tomli==2.1.0
typing_extensions==4.12.2
drf-spectacular==0.27.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-q==1.3.9
coveralls==4.0.1
django-dbml==0.7.0
uuid==1.30
Faker==33.3.0
django-json-widget==2.0.1
jsmin==3.0.1
tblib==3.0.0
geopandas==1.0.1
rasterio==1.4.3
rasterstats==0.20.0
pillow==11.1.0
topojson==1.9
matplotlib==3.9.4
